chore(remix-gallery): drop the stand-in card data now the feature ships
The `?remixdemo=` scaffolding existed because the treatment could not be judged
on real data — a few hundred images site-wide have an approved entry, so a
review pass saw a frame on nothing. It has served that purpose.

Removed: `demoRemixCount`, `demoRemixEntries`, `useRemixDemoDensity`, the
`demoActive`/`modulus` store fields and the hardcoded thumbnail list.

`REMIX_FRAME` and `useRemixPeelStore` stay — three card components import them —
so the module is renamed to `remix-card-state` rather than deleted, because a
file called `-demo` holding only production state is a trap for the next reader.

`useRemixCardData` collapses to reading the batched summary. This does NOT drop
the mod-only gate: it left with the demo branch, and the gate that matters lives
in `RemixGalleryBatchProvider`, which renders its children untouched and issues
no query when the flag is off, so a viewer without it reads no data.

Verified on real data rather than the stand-in: a host image with four approved
entries renders exactly one framed card on its owner's profile, and the fly-out
measures 4 of 5 hit-test points reachable — the same as the demo path scored.
